requirements/test.txt
# requirements/test.txt
-r common.txt
django-coverage==1.2.2
django-nose==0.1.3
mock==0.8.0
#nosexcover==1.0.7
#selenium==TODO
